Seperate command processor logic from IRC logic.
This moves the irc related stuff from the command processor to its own
package. It also seperates out irc spesfic stuff such as sending
messages to the network from the infomation about the command.

This makes it easier to add additional backends (xmpp, mattermost,
console) for simpler plugin testing and gives us the ability to support
other platforms (livecoding.tv chat, etc...).
