+++ src/main/java/uk/co/unitycoders/pircbotx/backends/irc/BasicMessage.java
@@ -16,7 +16,7 @@
  * You should have received a copy of the GNU General Public License along with
  * uc_pircbotx. If not, see <http://www.gnu.org/licenses/>.
  */
-package uk.co.unitycoders.pircbotx.commandprocessor.irc;
+package uk.co.unitycoders.pircbotx.backends.irc;
 
 import java.util.Iterator;
 import java.util.List;
@@ -25,6 +25,7 @@ import org.pircbotx.PircBotX;
 import org.pircbotx.User;
 import org.pircbotx.hooks.types.GenericMessageEvent;
 
+import uk.co.unitycoders.pircbotx.backends.AbstractMessage;
 import uk.co.unitycoders.pircbotx.commandprocessor.Message;
 
 /**
@@ -33,81 +34,17 @@ import uk.co.unitycoders.pircbotx.commandprocessor.Message;
  * An abstract version of the Bot Message interface which deals with some of
  * the common features which we can deal with for the child classes.
  */
-abstract class BasicMessage implements Message {
+abstract class BasicMessage extends AbstractMessage {
 	private final GenericMessageEvent event;
-	private final List<String> args;
 
 	public BasicMessage(GenericMessageEvent message, List<String> args) {
+		super(args);
 		this.event = message;
-		this.args = args;
-	}
-
-	public String getArguments() {
-		if (args.size() <= 2) {
-			return "";
-		}
-
-		// get the argument list
-		List<String> cmdArgs = args.subList(2, args.size());
-
-		//emulate String.join in java 1.7
-		StringBuilder argStr = new StringBuilder();
-		Iterator<String> argItr = cmdArgs.iterator();
-		while(argItr.hasNext()) {
-			argStr.append(argItr.next());
-			if (argItr.hasNext()) {
-				argStr.append(" ");
-			}
-		}
-
-		return argStr.toString();
-	}
-
-	public int getArgumentCount() {
-		return args.size();
-	}
-
-	@Override
-	public void insertArgument(int pos, String arg) {
-		args.add(pos, arg);
-	}
-
-	@Override
-	public String getArgument(int id) {
-		if (args == null || args.size() <= id){
-			throw new IllegalArgumentException("missing required argument");
-		}
-
-		return args.get(id);
-	}
-
-	@Override
-	public String getArgument(int id, String defaultValue) {
-		if (args == null || args.size() <= id){
-			return defaultValue;
-		}
-
-		return args.get(id);
 	}
 
 	@Override
 	public void respond(String response) {
 		event.respond(response);
-	}
-
-	@Override
-	public void respondSuccess() {
-		respond("The operation was successful.");
-	}
-
-	@Override
-	public String getMessage() {
-		return this.getArguments();
-	}
-
-	@Override
-	public  String getRawMessage() {
-		return event.getMessage();
 	}
 
 	@Override
